+++ src/orbital_engine/custom_types.py
@@ -1,30 +1,24 @@
-# from typing import NewType
+# Custom types for the orbital engine
 
 # Distances
-# Kilometers = NewType('Kilometers', float)
-# AstronomicalUnits = NewType('AstronomicalUnits', float)
 Kilometers = float
 AstronomicalUnits = float
 
+# Areas
+SquareMeters = float
+
 # Masses
-# Kilograms = NewType('Kilograms', float)
 Kilograms = float
 
 # Times
-# Seconds = NewType('Seconds', float)
 Seconds = float
 
-# EarthDays = NewType('EarthDays', float)
-# EarthYears = NewType('EarthYears', float)
 EarthDays = float
 EarthYears = float
 
 # Angles
-# Degrees = NewType('Degrees', float)
-# Radians = NewType('Radians', float)
 Degrees = float
 Radians = float
 
 # Velocities
-# KmPerSec = NewType('KmPerSec', float)
 KmPerSec = float